Refactored packages
 Changes to be committed:
	modified:   src/main/java/org/dimdev/dimdoors/block/entity/RiftBlockEntity.java
	modified:   src/main/java/org/dimdev/dimdoors/commands/CommandPocket.java
	modified:   src/main/java/org/dimdev/dimdoors/commands/CommandSaveSchem.java
	modified:   src/main/java/org/dimdev/dimdoors/pockets/PocketGenerator.java
	modified:   src/main/java/org/dimdev/dimdoors/pockets/PocketTemplate.java
	modified:   src/main/java/org/dimdev/dimdoors/rift/registry/RiftRegistry.java
	modified:   src/main/java/org/dimdev/dimdoors/rift/targets/EscapeTarget.java
	modified:   src/main/java/org/dimdev/dimdoors/rift/targets/PrivatePocketExitTarget.java
	modified:   src/main/java/org/dimdev/dimdoors/rift/targets/PrivatePocketTarget.java
	modified:   src/main/java/org/dimdev/dimdoors/rift/targets/PublicPocketTarget.java
	modified:   src/main/java/org/dimdev/dimdoors/rift/targets/RandomTarget.java
	modified:   src/main/java/org/dimdev/dimdoors/world/ModBiomes.java
	modified:   src/main/java/org/dimdev/dimdoors/world/ModDimensions.java
	renamed:    src/main/java/org/dimdev/dimdoors/world/pocketdimension/BlankBiome.java -> src/main/java/org/dimdev/dimdoors/world/pocket/BlankBiome.java
	renamed:    src/main/java/org/dimdev/pocketlib/BlankChunkGenerator.java -> src/main/java/org/dimdev/dimdoors/world/pocket/BlankChunkGenerator.java
	renamed:    src/main/java/org/dimdev/pocketlib/Pocket.java -> src/main/java/org/dimdev/dimdoors/world/pocket/Pocket.java
	renamed:    src/main/java/org/dimdev/pocketlib/PocketRegistry.java -> src/main/java/org/dimdev/dimdoors/world/pocket/PocketRegistry.java
	renamed:    src/main/java/org/dimdev/pocketlib/PrivatePocketData.java -> src/main/java/org/dimdev/dimdoors/world/pocket/PrivatePocketData.java
	renamed:    src/main/java/org/dimdev/pocketlib/VirtualLocation.java -> src/main/java/org/dimdev/dimdoors/world/pocket/VirtualLocation.java
	modified:   src/main/java/org/dimdev/gsonnbt/NbtReader.java
