+++ src/main/java/org/dimdev/dimdoors/commands/CommandPocket.java
@@ -16,7 +16,7 @@ import org.dimdev.dimdoors.pockets.SchematicHandler;
 import org.dimdev.dimdoors.rift.registry.RiftRegistry;
 import org.dimdev.dimdoors.rift.targets.EntityTarget;
 import org.dimdev.dimdoors.world.ModDimensions;
-import org.dimdev.pocketlib.Pocket;
+import org.dimdev.dimdoors.world.pocket.Pocket;
 import org.dimdev.util.Location;
 import org.dimdev.util.TeleportUtil;
