+++ src/main/java/org/dimdev/dimdoors/pockets/PocketTemplate.java
@@ -25,8 +25,8 @@ import org.dimdev.dimdoors.rift.registry.RiftRegistry;
 import org.dimdev.dimdoors.rift.targets.PocketEntranceMarker;
 import org.dimdev.dimdoors.rift.targets.PocketExitMarker;
 import org.dimdev.dimdoors.rift.targets.VirtualTarget;
-import org.dimdev.pocketlib.Pocket;
-import org.dimdev.pocketlib.PocketRegistry;
+import org.dimdev.dimdoors.world.pocket.Pocket;
+import org.dimdev.dimdoors.world.pocket.PocketRegistry;
 import org.dimdev.util.Location;
 import org.dimdev.util.math.MathUtil;
 import org.dimdev.util.schem.Schematic;
