+++ src/main/java/org/dimdev/dimdoors/rift/registry/RiftRegistry.java
@@ -8,9 +8,9 @@ import net.minecraft.world.PersistentState;
 import net.minecraft.world.World;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.dimdev.pocketlib.Pocket;
-import org.dimdev.pocketlib.PocketRegistry;
-import org.dimdev.pocketlib.PrivatePocketData;
+import org.dimdev.dimdoors.world.pocket.Pocket;
+import org.dimdev.dimdoors.world.pocket.PocketRegistry;
+import org.dimdev.dimdoors.world.pocket.PrivatePocketData;
 import org.dimdev.util.GraphUtils;
 import org.dimdev.util.Location;
 import org.jgrapht.graph.DefaultDirectedGraph;
