+++ src/main/java/org/dimdev/dimdoors/rift/targets/PublicPocketTarget.java
@@ -3,8 +3,8 @@ package org.dimdev.dimdoors.rift.targets;
 import net.minecraft.nbt.CompoundTag;
 import org.dimdev.dimdoors.pockets.PocketGenerator;
 import org.dimdev.dimdoors.rift.registry.RiftRegistry;
-import org.dimdev.pocketlib.Pocket;
-import org.dimdev.pocketlib.VirtualLocation;
+import org.dimdev.dimdoors.world.pocket.Pocket;
+import org.dimdev.dimdoors.world.pocket.VirtualLocation;
 import org.dimdev.util.Location;
 
 public class PublicPocketTarget extends RestoringTarget {
