+++ src/main/java/org/dimdev/dimdoors/world/ModDimensions.java
@@ -5,7 +5,7 @@ import net.minecraft.util.registry.Registry;
 import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.World;
 import net.minecraft.world.dimension.DimensionType;
-import org.dimdev.pocketlib.BlankChunkGenerator;
+import org.dimdev.dimdoors.world.pocket.BlankChunkGenerator;
 
 public final class ModDimensions {
     public static final RegistryKey<World> LIMBO = RegistryKey.of(Registry.DIMENSION, new Identifier("dimdoors:limbo"));
