+++ src/main/java/org/dimdev/dimdoors/world/pocket/Pocket.java
@@ -1,4 +1,4 @@
-package org.dimdev.pocketlib;
+package org.dimdev.dimdoors.world.pocket;
 
 import com.flowpowered.math.vector.Vector3i;
 import net.minecraft.entity.Entity;
